pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coo-assistant"
version = "0.1.0"
description = "AI-powered COO Assistant for internal operations management"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1",
    "anthropic>=0.40",
    "python-telegram-bot[all]>=21.0",
    "telethon>=1.36",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "structlog>=24.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
coo = "src.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100
